Add ProductRipository

+++ NadinsoftTask/Models/Product.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int ManufacturePhone { get; set; }
+        public short ManufacturePhone { get; set; }
         public string ManufactureEmail { get; set;}
         public bool IsAvailable { get; set; }
     }
